testing/compare: Fix image comparison RMS calculation.
The previous implementation did not compute the RMS error. It computed the RMS
in the difference of the number of colour components of each value. While this
computes 0 for equal images, it is incorrect in general. In particular, it does
not detect differences in images with the same pixels in different places. It
also cannot distinguish small changes in the colour of a pixel from large ones.